Add Cipher, Auth Algorithm Property into AvailableNetworkPack Class
