+++ ManagedNativeWifi/EnumDefinition/EncryptionType.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using static ManagedNativeWifi.Win32.NativeMethod;
 
 namespace ManagedNativeWifi
 {
@@ -55,7 +56,37 @@ namespace ManagedNativeWifi
 					return true;
 			}
 			encryption = default(EncryptionType);
-			return true;
+			return false;
 		}
+
+		public static bool TryConvert(DOT11_CIPHER_ALGORITHM source, out EncryptionType encryption)
+		{
+			switch (source)
+			{
+				case DOT11_CIPHER_ALGORITHM.DOT11_CIPHER_ALGO_NONE:
+					encryption = EncryptionType.None;
+					return true;
+				case DOT11_CIPHER_ALGORITHM.DOT11_CIPHER_ALGO_WEP40:
+					encryption = EncryptionType.WEP;
+					return true;
+				case DOT11_CIPHER_ALGORITHM.DOT11_CIPHER_ALGO_TKIP:
+					encryption = EncryptionType.TKIP;
+					return true;
+				case DOT11_CIPHER_ALGORITHM.DOT11_CIPHER_ALGO_CCMP:
+					encryption = EncryptionType.AES;
+					return true;
+				case DOT11_CIPHER_ALGORITHM.DOT11_CIPHER_ALGO_WEP104:
+					encryption = EncryptionType.WEP;
+					return true;
+				case DOT11_CIPHER_ALGORITHM.DOT11_CIPHER_ALGO_WPA_USE_GROUP:
+					encryption = EncryptionType.None;
+					return true;
+				case DOT11_CIPHER_ALGORITHM.DOT11_CIPHER_ALGO_WEP:
+					encryption = EncryptionType.WEP;
+					return true;
+			}
+			encryption = default(EncryptionType);
+			return false;
+        }
 	}
 }
